:wrench: chore: cria função auxiliar para cores
